requirements.txt
flask==2.0.1
werkzeug==2.0.1
mysql-connector-python==8.0.27
numpy==1.20.3
pandas==1.3.4
matplotlib==3.5.0
seaborn==0.11.2
